Authentication

+++ src/main/java/com/backend/chatop4/repository/RentalRepository.java
@@ -4,7 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.backend.chatop4.model.Rental;
 
-public interface RentalRepository extends JpaRepository<Rental, Long> {
-  
-  
+public interface RentalRepository extends JpaRepository<Rental, Integer> {
+
 }
